common.py: Change 'device-tree' repo update process used for TC5 images
In update_dt_git_repo(), the code was incorrectly passing a GitPython
Head object as the refspec argument in the Remote.fetch() method, which
should be a string or list of strings.

Newer versions of GitPython added a check that verifies if refspec is of
the expected type, throwing an error if that isn't the case. Because of
this the 'dt checkout --update' command fails if TCB was built with recent
versions of the GitPython library.

This commit removes the unnecessary refspec argument.

Signed-off-by: Lucas Akira Morishita <lucas.morishita@toradex.com>