tests/integration: Workaround to Aktualizr reporting "no updates found" after TCB deploy
When the TCB performs a deploy, the OSTree version changes, but it does not
cause a change in the version recognized by Aktualizr. This leads to a
situation where, if an update is requested for the version before the TCB
deploy, Aktualizr reports "no updates found" and does not perform the update,
as it believes the requested update is for a version already present on the
device. The workaround used is to request an update to a nightly if
using a release and vice-versa.

Related-to: TOR-3518
Related-to: TOR-490
Related-to: TCCP-864

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>