backend/platform: fetch latest director root.json on lockbox download
The director service in torizon cloud will ensure the root.json expires
later than all offline-metadata as part of servicing the request to
GET:api/v1/admin/repo/root.json. If necessary, the director will create
a new version of root.json with a later expiration.

It is imperative to trigger this validation during the creation of a
lockbox to prevent issues where a user cannot install a valid lockbox
because the root metadata has expired.

Fetching the root metadata directly with the version (ie: 1.root.json)
will not trigger this validity check, and therefore must be preceeded
by fetching the naked root.json

Signed-off-by: Ben Clouser <dev@benclouser.com>