bundle: Add support for passing environment variables to DIND
Add a new parameter to the bundle command, namely "--dind-env" which
allows one to set environment variables to be passed to the DIND
(Docker-in-Docker) instance that TorizonCore Builder employs to fetch
container bundles. Since DIND executes a Docker Daemon instance inside
itself, that daemon will see those variables which in turn allows for
example a proxy server to be set and utilized when fetching container
bundles which was a missing feature before this implementation.

For setting the proxy, users would pass to the bundle command parameters
such as below (variables are described in the Docker documentation):

* --dind-env HTTP_PROXY="http://<http-proxy-ip>:<proxy-port>"
* --dind-env HTTPS_PROXY="http://<https-proxy-ip>:<proxy-port>"
* --dind-env NO_PROXY="..."

Related-to: TOR-3662

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>