build: Add support for passing parameters and env vars to DIND
Add new properties "dind-params" and "dind-env" to the build file to
allow users to specify extra parameters and environment variable values
to be passed to the DIND (Docker-in-Docker) instance that TorizonCore
Builder employs to fetch container bundles. This is similar to what was
done for the bundle and platform commands in previous commits.

Related-to: TOR-3662

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>