tests/integration: Refactor get_tezi_images to download latest versioned images
Previously, get_tezi_images was downloading specific images, that had
the versions hardcoded on script. This commit changes this to download
the latest release, monthly or nightly.
Also, it's possible to define the YOCTO_BRANCH to request a specific
version of Torizon, e.g. 'kirkstone-6.x.y' and 'scarthgap-7.x.y'.
Finally, download only the image for the device tested, instead of
downloading images for all devices.

Related-to: AVAL-18

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>