tests/integration: Add --ignore-different-secondaries-between-updates to AVAL
A new secondary for fuse was added to Torizon. Because of this, when AVAL
tries this update path:

(image with secondary) -> (image withouth secondary) -> (image with secondary)

Aktualizr/Uptane breaks due to an artificial limitation imposed by the platform
to prevent security issues. Thus we must always make sure to remove secondaries
that are not in the intersection between the images. In the current case, this
is only the `fuses` secondary.
This flag allow AVAL to delete /var/sota/storage/fuse path, as a workaround
for this.

Related-to: TCB-505
Related-to: TOR-3635

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>