tests/integration: Enable tests that access localhost network in CI
These tests were failing because the container running the jobs in CI does not
have the --network=host configuration. Therefore, it couldn't access the host's
network. To solve this, the commands that tried to download files from servers
created with torizoncore-builder were adapted to run in a busybox container
with the --network=host configuration.

Related-to: TCB-503

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>