aval-tests: Add variable to allow tests to fail plus improvements
Add pipeline variable AVAL_ALLOW_FAILURE (default value: false) which
one can set to true to allow the pipeline to continue running when AVAL
tests fail. This can be useful to deal with instabilities of the
infrastructure.

Also, do not rely on the pipeline source being set to "schedule" to
enable the AVAL-related jobs. Relying on the source makes it harder for
testing the pipeline. If someone does not want the AVAL-related jobs to
be present on the pipeline they can always set AVAL_DISABLED=true.

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>