tests/integration: Enable "images serve: check zeroconf" test in CI
To run the avahi-browse-domains command, it was necessary to use an Alpine
container, which configures DBUS, starts the avahi-daemon in the background,
and finally runs avahi-browse-domains.

Additionally, when running the tests with dind configured to use a TCP port (as
is the case with this repo), Avahi was unable to discover services on the
network with --network=host. On the other hand, when dind is mounted using
docker.sock, Avahi was able to find network services, indicating that there is
a conflict  between dind running with the TCP port and Avahi. To solve this, a
Docker network tcb_network was created, where the command torizoncore-builder
images serve is first run in background, and then avahi-browse-domains is
executed, making it possible to discover services on the network in this
Docker network.

Related-to: TCB-503

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>