aval-tests: Change the workaround for mismatch between database and OSTree
The previous workaround was generating a problematic edge case. Let’s suppose
the existence of two parallel tests, A and B. With this previous workaround,
test A runs the TCB tests, releases the device, and then takes it again to
update it to a different version, aiming to sync the Aktualizr’s database with
OSTree. However, if test B takes the device after A’s TCB tests but before A’s
workaround update and attempts an update to the Torizon version before A’s TCB
tests, the device will fall into the same mismatch bug.

Because of this, the workaround was changed to delete the Aktualizr’s database,
which forces Aktualizr to recreate it, this time based on OSTree, bringing both
into sync again. With this new workaround, the device is not released until the
workaround is completed, there is no risk of encountering these edge cases, and
it also speeds up testing by saving one update.

Related-to: TOR-3518
Related-to: AVAL-53
Related-to: TCB-490

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>