tests/integration: Implement JUnit test report
This implementation of the JUnit report in this repository is part of a series
of implementations across various Toradex repositories, aiming to standardize
test reporting and facilitate the aggregation of test results for all Torizon
projects.
For this, 'test-torizoncore-common' was split into two executions, one for
Intel and one for Raspberry, so each pipeline runs only one test execution.
In addition to speeding up tests by increasing parallelism, the JUnit report
becomes simpler when only a single BATS test results file is provided.

Related-to: IN-11
Related-to: TCB-506

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>