tests/integration: push aval-docker container to CI_REGISTRY
The step of push container to CI_REGISTRY was missing. Then, the job was
using the last aval-docker container available in container registry, that is
deprecated related to upstream of AVAL.
Also, use GITLAB_DOCKERREGISTRY_SUFFIX to tag the container, that
increase reliability in cases when a lot of pipelines are running in
parallel, and all of them were pushing using main tag.

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>