tests/integration: Implement XRay for jobs running in nightly of TOS7
This change adds support for Xray report uploads to Jira, enabling better
visibility into the current status of TCB with the latest nightly builds of
Torizon. Since the goal is to reflect the integration state with the most
recent system image, the reporting is limited to nightly test jobs only.

Related-to: IN-11

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>