platform: Handle cases where an image is specified by tag+digest
In theory, specifying a Docker image by a tag plus a digest (such as in
`image:tag@sha256:<hex-digest`) doesn't make much sense, but Docker does
accept it by basically ignoring the tag. However, the implementation in
TorizonCore Builder did not work in such cases. With the present commit
we try to mimic Docker's behavior by also ignoring the tag. This has an
effect both when canonicalizing a compose file (via `platform push`) and
when generating Lockboxes (via `platform lockbox`).

Related-to: TCB-689

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>