secboot: Add new sign-kernel and sign-bootloader-hab commands
Add new commands that re-signs existing signed Torizon OS 7 images.
'secboot sign-kernel' and 'secboot sign-bootloader-hab' will sign the
kernel fitImage and the bootloader container (flash.bin/imx-boot) for
HAB-compatible devices, respectively.

For each command the end result is a copy of the input OS image with the
aforementioned components re-signed.

For now the only supported device for both commands is the Verdin i.MX8M
Plus. Other boards are planned to be supported in the future.

The kernel fitImage is signed with a user-provided directory with the
RSA key and certificate. The bootloader container is signed with a
directory tree with keys, certificates, SRK table and SRK fuse hash
generated beforehand using the NXP Code Signing Tool.

Signed-off-by: Lucas Akira Morishita <lucas.morishita@toradex.com>