platform: Only fetch Docker manifests for the needed platforms
Originally TorizonCore Builder was fetching all manifests associated
with each Docker image referenced by a Lockbox. This caused unnecessary
requests to the registry in turn using part of the requests quota set by
some registries (particularly from Docker Hub one).

With this commit we only fetch the manifests of the images for the
needed platforms (the ones specified by the user with the --platform
switch to `platform lockbox` or the default ones, if not specified).

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>