platform: Generate Lockbox tarballs via skopeo
The overall goal here is to produce image tarballs in a format that is
independent of the container engine used to fetch them. For that,
instead of running "docker save" to produce the tarballs we switch to
using the skopeo tool. Unfortunately, out of the box tarballs produced
by skopeo are also not fully compatible with the validation code in
Aktualizr, but by making some amendments (related to image canonica-
lization) to them we can ensure compatibility with the current
validation code. The required amendments are also part of this commit.

Besides the above, also apply a workaround related to the size of the
produced tarballs (TOR-3854).

With this commit, the image generation via skopeo becomes the default
but this can be changed by setting the environment variable
"LOCKBOX_TARBALL_GENERATOR" which can be set to "skopeo" or "docker".

If set to "skopeo", the application of the canonicalization workaround
can be controlled by the env. variable "APPLY_SKOPEO_CANON_WKAROUND".

The other workaround, which is relevant whether "skopeo" or "docker" is
selected for image generation is controlled by the environment variable
"APPLY_AKTUALIZR_BLKREAD_WKAROUND".

Nevertheless, users are not expected to have to change any of these
variables.

Related-to: TCB-692, TOR-3854

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>