tests: Update platform tests according to latest behavioral changes
Make the "platform lockbox" tests more granular and adapt them to the
latest changes in behavior. The changes on the tests were needed due to
a new check on TorizonCore Builder which prevents the same image from
being referenced multiple times with different names. Though this is not
expected to happen in practice, some of the tests fell exactly into that
case.

Besides updating the tests, also create a new folder containing a script
to populate the platform account with the required compose files for the
updated platform tests plus the compose files themselves. The correspon-
ding Lockboxes still need to be created manually though, which is a
situation that can be improved in the future.

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>