tests/integration: Run the possible jobs in cloud runners
This commit updates the pipeline configuration to run jobs that don't require
internal resources in cloud runners.

Jobs related to "torizoncore" and "aval" continue to run on internal runners:
- "torizoncore" jobs require access to the internal Artifactory server.
- "aval" jobs establish connections to modules over the internal network.

Also, add WSL tests using AVAL to run together with Linux tests in push
pipelines. This was done because it wasn’t necessary to revalidate all the
jobs on WSL, and several jobs ended up being duplicated in the previous
approach of selecting a runner tag via environment variable and running the
entire pipeline.
Moreover, this method of dynamically defining the runner tag through an
environment variable is not supported by GitLab, which led to several issues
when it became necessary to make the code more complex to support the Cloud
runners that use specific tags.

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>