tests/integration: Add default parameters to get_images scripts.
This gives two options to the user: (1) run the script with default
values without set any variable; (2) customize enviroment variables if needed.

The change was made to avoid using the CONFIGME approach that forces the
user to do a change in the script.

The README is also updated to reflect the new script usage.