gitlab-ci: Disable linter jobs on Jenkins pipelines
The pipelines triggered by Jenkins are purely for integration purposes, aiming
to identify new bugs in the torizoncore-builder when new versions of Torizon
are released. They do not contain any changes to the torizoncore-builder code.
Therefore, since there are no code changes, it doesn't make sense to run the
linter in these pipelines, as it would be just a waste of resources.
This commit adds a rule that disable linter pipelines when the source is
a trigger token.

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>