gitlab-ci: Report results with XRay even if a previous step fails
After commit e37b1f2, the GitLab jobs started to fail if any test fails,
instead of failing only when they don't reach the end.

Therefore, after this commit, if a previous test stage failed, the
aggregate-reports job would not run and the error wouldn't be reported. With
this commit, it now runs regardless of previous steps failing.

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>