gitlab-ci: Use customized GitLab SaaS runners for CI jobs
This change to the GitLab runners is a request from the Automation team, which
is now adopting the GitLab runners as the default runners to be used by the
project. A study was carried out to determine which GitLab runner is the best
fit for TCB. In this study, a full AVAL pipeline was executed using each of
GitLab’s SaaS-linux runners, from small up to 2xlarge.

The use of these customized SaaS runners aims to minimize the time of each
stage, up to the point where the cost becomes higher than the performance gain.
These changes are being made in this commit:

- default: use saas-linux-small-amd64, as it has the lowest cost factor.

- lint-python and build-amd64 jobs: use saas-linux-large-amd64, because the job
time decreased significantly up to the large runner. Beyond the large runner,
the costs increased significantly and did not justify the gain in performance.

Related-to: TCB-733

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>