isolate, deploy: Use Fabric library to run shell commands remotely
This fixes an IO desynchronization issue when sending remote sudo
commands that could happen if the device has Torizon OS 7.3.0 or higher
installed: the password would be sent to the device before sudo is ready
to receive it, which would cause sudo to hang waiting for the password
until its default timeout (5 minutes) gets reached and fail.

This wasn't reproducible on previous Torizon OS 7 versions. 7.3.0
introduced a newer version of sudo (1.9.17p1), which could explain
this difference in behavior: probably this version takes a bit longer to
show the password prompt, just enough for the issue to show up.

One way avoid this problem is to send shell commands remotely using a
library designed to do this task, like Fabric. It can detect the sudo
password prompt and auto-respond to it accordingly, which is a more
robust method than the previous one.

Related-to: TCB-729

Signed-off-by: Lucas Akira Morishita <lucas.morishita@toradex.com>