tests/integration: Return error if test execution key is not set
Test execution key variable is a requirement for that script to be successful,
however the job is returning only a warning.

This commit changes this behavior to return an error if this variable is
not set, which indicates better that something needs to be fixed, in
that case defining this variable.

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>