secboot: Refactor commands to affect the current unpacked image
Refactor sign-kernel and sign-bootloader-hab to affect the current
unpacked image instead of asking for an input image. This way the
secboot commands would follow the stateful workflow applied for other
standalone commands, which is similar to this:

images unpack -> standalone commands -> union -> deploy

This workflow also guarantees that the re-signed kernel FIT is included
in an OSTree commit and is properly tracked, which wasn't the case with
the previous implementation of the 'sign-kernel' command.

The sign-kernel command now gets the input kernel FIT directly from the
current rootfs deployment of the unpacked image instead of expecting the
FIT to be in the tarfile with the signing binaries
(tcb_signing_files.tar.gz).

As the bootloader container, the tarfile with the signing binaries, and
the fusing instructions text file are not tracked by OSTree, the new
implementation of sign-bootloader-hab internally associates those files
with an OSTree commit when running the union command. This was done to
make sure that running a deploy command to a local directory always
outputs the correct content for a given union branch, even for files not
tracked by OSTree.

Running a remote deploy to a device will NOT apply the signed
bootloader, as it is essentially an OSTree-only update, and the
bootloader binary is not tracked by OSTree.

Now when running images unpack on an image with the signing binaries
tarball, both the bootloader and the tarfile are internally tracked with
the current OSTree commit of the unpacked image.

Related-to: TCB-731

Signed-off-by: Lucas Akira Morishita <lucas.morishita@toradex.com>