build: Add initial support to sign OS components of supported images
Integrate the features of the 'secboot sign-kernel' and 'secboot
sign-bootloader-hab' commands into the 'build' command, allowing users
to sign the kernel FIT and the bootloader container of supported Torizon
OS images using the stateless workflow of the build command.

The tcbuild.yaml template has been updated with the new fields,
alongside brief explanations for each one.

Related-to: TCB-731

Signed-off-by: Lucas Akira Morishita <lucas.morishita@toradex.com>