tests/integration: Remove docker network usage in zeroconf test
First of all, this approach of creating a Docker network does not reflect the
behavior of the user when running this command, since the torizoncore-builder
commands do not use Docker networks, but instead use "--network=host" in
Linux.

In addition, this Docker network is not necessary for these tests; they succeed
in Linux just fine using the standard approach of passing "--network=host" to
the torizoncore-builder Docker command. It fails in WSL, but images
serve commands are not suported in WSL.

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>