aval-tests: Run only one AVAL job for the same machine at the same time
Previously, all AVAL jobs ran in parallel. For each machine, there are usually
three jobs:
- Kirkstone, release
- Scarthgap, release
- Scarthgap, nightly

However, in practice, there is usually only one machine of each type in AVAL.
As a result, one of these three jobs will run while the other two wait to
acquire the machine, retrying and occupying runners, which can even result in
a timeout if the machine cannot be locked. This is the reason for the error
"Wasn't able to lock the device <uuid> after 80 attempts".

This commit makes use of 'resource_group' so that jobs for the same machine
run sequentially, ensuring that only one job at a time acquires the device
in AVAL and avoiding lock contention across jobs in the same pipeline.

Related-to: TCB-737

Signed-off-by: Lucas Bernardes <lucas.bernardes@toradex.com>