backend/isolate: Support filenames with leading or trailing spaces
This fixes a (very) particular case when the device /etc has files or
directories with leading or trailing whitespaces in their names.

Now any whitespace in .tcattr filename entries is replaced by a
backslash followed by its respective 3-digit octal ASCII code (\040).

Previously the resulting .tcattr had filenames without any changes i.e.
with the exact number of whitespaces. However when running the union
command, setfacl ignores any leading/trailing whitespaces in the
file/directory name entry, resulting in it looking for filenames
without them.

This was originally detected in our test pipeline, where the device
under test had files with this name pattern in /etc, probably as a
result of a previous 'isolate' test suite that created them, but never
deleted them as the it failed and the cleanup routine ended up not
being executed.

Signed-off-by: Lucas Akira Morishita <lucas.morishita@toradex.com>