Throw error when running unsupported features for kernel in FIT format
When running certain commands/features, TorizonCore Builder will now
first check if the unpacked image has the kernel in FIT format, and
abort execution if the current implementation doesn't support it.

This applies to dt, dto, kernel and splash commands, as well as their
respective features in the build command.

Related-to: TCB-755

Signed-off-by: Lucas Akira Morishita <lucas.morishita@toradex.com>