tests/integration: Update tests to check for FIT kernel support
Update some tests to first check if the OS image being tested has
the kernel in FIT format, as some customization features (i.e. dt, dto,
kernel, splash commands) are not currently available in this case. The
tests will be skipped if a FIT kernel is detected.

Also add FIT-only test cases to check if an error message is properly
being shown for these unsupported features.

Related-to: TCB-755

Signed-off-by: Lucas Akira Morishita <lucas.morishita@toradex.com>