tests: Use a single directory to get both signed/unsigned images
Previously the test workflow looked at a separate directory to get a
signed input image that can be used with the secboot command.

Since the tests can now detect whether the Torizon OS image to be tested
has a FIT kernel or not, we don't need to separate signed and unsigned
images: it is only necessary to check for a FIT kernel and if the image
has the tarball with the signing artifacts.

Signed-off-by: Lucas Akira Morishita <lucas.morishita@toradex.com>