dt: Add support for dt operations on kernel FIT images
Add support for dt operations (status, apply) on images having the
kernel in FIT format.

Users of standalone commands will realize the behavior of TCB is a bit
different when the tool determines the kernel is in FIT format.

First of all, the changes made by command "dt" will go to the "kernel"
directory in the storage (unlike with non-FIT kernels where changes go
to the "dt" directory). This means users would need to pass the "kernel"
directory to the "union" command when generating the OSTree commit with
their changes. The reason for this change in behavior is that "dt apply"
operation impacts the kernel (which will actually hold the contents of
the applied DTB) and a few other artifacts (uEnv.txt, overlays.txt) but
all these changes only make sense when made together. By putting them
all into the same changes directory (namely, the "kernel" dir) we ensure
users will always apply them at the same time (atomically).

A second user visible difference is that, with non-FIT the "dt apply"
command deletes all DT/DTO changes in storage before actually applying
new changes (one could argue that this is a bug but it's best to keep it
for compatibility reasons). This means a previously applied DT is thrown
away when "dt apply" is executed. With FIT, applied DTs actually
accumulate inside the to-be-deployed kernel which is more in line with
the statefulness of the tool (when standalone commands are employed).

The above behavioral differences are mostly transparent to users of the
"build" command.

Related-to: TCB-761

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>