dt: Change how an updated uEnv.txt is generated
Originally the "dt apply" command took the uEnv.txt file always from the
base commit. This means, if other commands made changes / customizations
to that file (which would be inside a changes directory), they would be
lost upon the execution of "dt apply". Though this was wrong in concept,
it didn't cause issues because no other commands actually touched the
uEnv.txt file. However, this will soon change with the new method for
letting TCB set arguments to be passed to the kernel without the need of
an overlay (the current method).

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>