dto: Add support for dto operations on kernel FIT images
Add support for dto operations ("apply", "list", "status", "remove") on
images having the kernel in FIT format; the "deploy" operation, which
wasn't supported since Torizon OS 6, it's also not supported with FIT
images.

Similarly to what we have with the "dt" command, the behavior of TCB
regarding the "dto" command is a bit different when the tool determines
the kernel is in FIT format.

The main difference relates to the directory where the changes applied
by the command are kept inside the storage: they all go to the "kernel"
directory. The reason for this is that the dto operations "apply" and
"remove" impact the kernel (which will actually hold the contents of the
applied DTB) and a few other artifacts (uEnv.txt, overlays.txt) but all
these changes only make sense when made together. By putting them all
into the same changes directory (namely, the "kernel" dir) we ensure
users will always apply them at the same time (atomically).

Apart from the above, the overall behavior of the various commands is
kept the same as with FIT. For example, to mimick the non-FIT case, when
an overlay that existed in the base image is deleted, it is actually
restored from the base image (copied from the original kernel) instead
of simply deleted from the kernel being customized.

The behavioral differences are expected to be transparent to users of
the "build" command.

Related-to: TCB-762
Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>