kernel: Add support for custom bootargs ops on kernel FIT images
Add support for custom boot arguments set/get/clear operations on images
having the kernel in FIT format.

When TCB detects that the image has support for the new argument passing
method (based on variables in uEnv.txt), it automatically selects it;
otherwise it falls back to the old method (based on an overlay).

Besides setting the custom arguments, when TCB finds the Secure Boot
kernel arguments overlay inside the FIT image (containing all arguments
for booting the system, not only the custom arguments), it also updates
it with the custom arguments to satisfy the Kernel command line
protection feature of Torizon OS.

The implementation in this commit required an update of pylibfdt to
solve a bug where an exception was not thrown when reading a property
not present in the FDT.

Related-to: TCB-763
Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>