kernel: Refactor the build_module command implementation
Refactor the implementation of the kernel build_module command, both at
the cli and backend levels. This is done in preparation for adding
support for FIT kernels.

In addition to the refactoring (which does not change behavior), a minor
update was made to require the SRC_DIR parameter when running the
command. Previously, this parameter was not enforced by the command-line
parser, even though it was required internally. As a result, invoking
the command without a source directory would lead to an error at
runtime.

Related-to: TCB-763
Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>