kernel: Fix bug handling the --autoload switch of build_module
If kernel build_module was executed multiple times for a given module
with the switch --autoload being passed, the file listing which modules
to automatically load (under /etc/modules-load.d) would accumulate
occurrences of the same module name.

The present commit solves the situation by deleting any occurrences of
the module name from the file, before adding it at the end. This allows
users to fix the module loading order if needed.

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>