build: Avoid duplicated error prefix during "build"
Some message strings embedded into exceptions thrown by TorizonCore
Builder have an "Error:" prefix and when those exceptions are cast
during the execution of the "build" command, that command also adds the
same prefix to the final message. As a result, the final message shown
to the user contains a duplicated error prefix. E.g.:

Error: Error: the Torizon OS image you are customizing has a kernel in \
       FIT format but it does not support the uEnv method...

This commit solves this issue by only adding the prefix if it is not yet
present in the original message.

Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>