kernel: Prepend custom kargs with secboot images
When detecting that the image has secboot features, we now
prepend (rather than append) the custom kargs string to the existing
bootargs. This creates a limitation in that users won't be able to
override bootargs explicitly set in the base image, but it makes the
change in bootargs compatible with the kernel command-line protection
feature (which is part of the Toradex hardening features in
U-Boot). This relies on the new variables torizon_bootargs_{l,r}
provided by the latest boot script (uEnv.txt).

If the image is not a secboot one, the kargs are still appended as
usual. This is to keep the behavior exactly as before for non-secboot
users, avoiding a breaking change from their perspective.

Related-to: TCB-763
Signed-off-by: Rogerio Guerra Borin <rogerio.borin@toradex.com>