tests: Fix DTO test when running TCB image with 'early-access' tag
Fix a test that only fails when running a TorizonCore Builder image with
the 'early-access' tag.

This specific tag causes an additional message to be printed to stderr
when running every command informing that the image is an early-access
version. This extra message made the test fail as it only expected the
one line of output from the 'dto status' command itself.

Also increase the verbosity level for a WIC build test to facilitate
debugging.

Signed-off-by: Lucas Akira Morishita <lucas.morishita@toradex.com>